// src/client/response/payload/payment/payment-method.response.ts
export interface PaymentMethodResponse {
    id: string;
    type: string;
    name: string;
    currency: string;
    primary_buy: boolean;
    primary_sell: boolean;
    allow_buy: boolean;
    allow_sell: boolean;
    allow_deposit: boolean;
    allow_withdraw: boolean;
    instant_buy: boolean;
    instant_sell: boolean;
    created_at: string;
    updated_at: string;
    resource: string;
    resource_path: string;
}